# swagger/swagger.yml
openapi: 3.0.1
info:
  title: Yo!Photo 👻
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
  - url: http://localhost:8000/api/v1
  - url: http://localhost:8000/api/v2
...
paths:
  /users:
    get:
      tags:
        - User
      summary: Get all user
      description: >
        <img src="https://www.prisma.io/docs/static/8cbb3ea8b5fc961b73cd2e583313f3a9/a6d36/offset-skip-take.png" alt="drawing" width="400"/>

      operationId: getMany
      parameters:
        - name: skip
          in: query
          description: skip
          required: false
          schema:
            type: integer
            example: 0
        - name: take
          in: query
          description: take
          required: false
          schema:
            type: integer
            example: 50
            minimum: 0
            maximum: 100
      responses:
        "200":
          description: Return an array of user objects
          content:
            application/json:
              schema:
                type: array
                # items:
                # type: object
                # $ref: "#/components/schemas/User"
                $ref: "#/components/schemas/GetAllUsers200"
...
  /users/profile/:
    get:
      tags:
        - Profile
      summary: Get all profiles
      parameters:
        - name: skip
          in: query
          description: skip
          required: false
          schema:
            type: integer
            example: 0
        - name: take
          in: query
          description: take
          required: false
          schema:
            type: integer
            example: 100
      responses:
        default:
          description: successful operation
          content: {}
      x-codegen-request-body-name: body
    put:
      tags:
        - Profile
      summary: Update users own profile
      requestBody:
        required: false
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                  example: 4
                name:
                  type: string
                  example: Arman Maulalan
      responses:
        default:
          description: successful operation
          content: {}
      x-codegen-request-body-name: body
  /users/profile/userid/{id}:
    get:
      tags:
        - Profile
      summary: Gets profile by userId
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: string
            example: 4b6ddf52-40f9-4545-ab3a-bdd6eb44746b
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /users/profile/username/{username}:
    get:
      tags:
        - Profile
      summary: Gets specific profile by Username
      parameters:
        - name: username
          in: path
          description: User name
          required: true
          schema:
            type: string
            example: alice
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /photos:
    get:
      tags:
        - Photo
      summary: Get all photos from database with limit
      parameters:
        - name: limit
          in: query
          description: limit
          required: true
          schema:
            type: integer
            example: 50
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: integer
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        type: object
                      $ref: "#/components/schemas/Photo"
                      additionalProperties:
                        $ref: "#/components/schemas/PhotoDetail"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Photo
      summary: Update photo detail by photo Id
      requestBody:
        required: false
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                  example: 0050580b-1a3e-452e-adb6-7618f36519a9
                title:
                  type: string
                  example: This is a new title
                description:
                  type: string
                  example: This should be a very long description
                tag:
                  type: string
                  format: csv
                  example: robot,kuda
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
  /photos/user/{id}:
    get:
      tags:
        - Photo
      summary: Get all photos from a specific user
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
  /photos/{photoId}:
    get:
      tags:
        - Photo
      summary: Get a specific photo by photo ID
      parameters:
        - name: photoId
          in: path
          description: photo ID
          required: true
          schema:
            type: string
            format: uuid
            example: 277dfc20-a881-4105-b344-1dfca911166b
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
  /photos/upload:
    post:
      tags:
        - Photo
      summary: Upload photo to collection User
      description: Logged user only
      operationId: uploadePhoto
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo:
                  type: array
                  items:
                    type: string
                    format: binary
                  oneOf:
                    - minItems: 1
                      maxItems: 1
                    - minItems: 3
                      maxItems: 3
      responses:
        default:
          description: successful operation
          content: {}
      x-codegen-request-body-name: body
...
components:
...
  schemas:
    GetAllUsers200:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          type: string
          example: Success
        result:
          # example:
          # $ref: "#/components/schemas/ArrayOfUsers"
          type: array
          items:
            type: object
            anyOf:
              - $ref: "#/components/schemas/User"
              - $ref: "#/components/schemas/User"
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 7f3ed61f-cd69-44c4-8a6e-fd723ff896ec
        userName:
          type: string
          example: kharefoot7
        email:
          type: string
          format: email
          example: mfudger7@hatena.ne.jp
        password:
          type: string
          example: $2a$10$71IM5uVMR5KE1skbUySVM.6ONjZ9wj1Rxq08FNRQjXiJSykqR445O
        role:
          type: string
          example: USER
        isActive:
          type: boolean
          example: true
        lastLoginAt:
          type: string
          format: date-time
          example: 2022-03-06T23:16:15.850Z
        createdAt:
          type: string
          format: date-time
          example: 2022-03-03T19:38:30.610Z
        updatedAt:
          type: string
          format: date-time
          example: 2020-01-16T12:43:16.000Z
...
    Profile:
      type: object
      title: Profile
      properties:
        userId:
          type: string
          format: uuid
          example: 7f3ed61f-cd69-44c4-8a6e-fd723ff896ec
        name:
          type: string
          example: Harley Gaffer
        addres:
          type: string
          example: 00 Fremont Pass
        profilePhoto:
          type: string
          format: url
          example: https://randomuser.me/api/portraits/men/27.jpg
        coverPhoto:
          type: string
          format: url
          example: https://picsum.photos/id/8/800/600
        createdAt:
          type: string
          format: date-time
          example: 2020-03-28T11:37:03.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2020-03-28T13:48:38.000Z
...
    Photo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 277dfc20-a881-4105-b344-1dfca911166b
        title:
          type: string
          example: Amblyrhynchus cristatus
        description:
          type: string
          example: Team-oriented real-time intranet
        path:
          type: string
          format: url
          example: https://picsum.photos/seed/28/800/600
        userId:
          type: string
          format: uuid
          example: 7f3ed61f-cd69-44c4-8a6e-fd723ff896ec
        createdAt:
          type: string
          format: date-time
          example: 2017-12-06T00:00:00.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2017-12-06T00:00:00.000Z
    PhotoDetail:
      type: object
      properties:
        photoId:
          type: string
          format: uuid
          example: 277dfc20-a881-4105-b344-1dfca911166b
        fileName:
          type: string
          example: 260d3b01-129c-4284-99a2-78a403041e9e
        originalName:
          type: string
          example: Venenatis.png
        mimeType:
          type: string
          example: image/png
        encoding:
          type: string
          example: base64
        size:
          type: integer
          example: 3390164
        isPrivate:
          type: boolean
          example: false
        views:
          type: integer
          example: 427
        likes:
          type: integer
          example: 14
        downloaded:
          type: integer
          example: 13
        starRating:
          type: integer
          example: 5
        cameraMake:
          type: string
          example: Ricoh
        cameraModel:
          type: string
          example: Hasselblad
        shutterSpeed:
          type: string
          example: 1/500
        aperture:
          type: string
          example: 35 mm
        focalLength:
          type: string
          example: 35 mm
        iso:
          type: integer
          example: 800
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
...